[Numerics] fix build ...

+++ Numerics/src/NonSmoothSolvers/FrictionContact/FrictionContact3D_AVI_gams.c
@@ -133,7 +133,7 @@ static void setDashedOptions(const char* optName, const char* optValue, const ch
   }
 }
 
-static int SN_rm_normal_part(int i, int j, double val, void* env)
+static ptrdiff_t SN_rm_normal_part(ptrdiff_t i, ptrdiff_t j, double val, void* env)
 {
   if (i%3 == 0)
   {
